fix: fixed overflow on integer in reward query

+++ backend/Application/Api/GraphQL/Metrics/RewardMetricsQuery.cs
@@ -31,7 +31,7 @@ public class RewardMetricsQuery
                     where time between @FromTime and @ToTime;";
         var data = await conn.QuerySingleAsync(sql, queryParams);
         
-        var sumAmount = (int)data.sum_amount;
+        var sumAmount = (long)data.sum_amount;
 
         var bucketParams = queryParams with
         {
